requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
numpy>=1.24.0 
flask
cloudinary
torch
torchvision
deep-sort-realtime
paddleocr  
mysql-connector-python  